# Wrestling/openapi.yaml
openapi: 3.0.3
info:
  title: Wrestling API
  version: 5.0.0
  description: REST API for the Wrestling catalog, content and progress-ready frontend.
servers:
  - url: /api
paths:
  /health:
    get:
      summary: Health check
      responses:
        '200': {description: Healthy}
  /ready:
    get:
      summary: Readiness check
      responses:
        '200': {description: Ready}
  /elements:
    get:
      summary: List and filter workout elements
      parameters:
        - {name: category, in: query, schema: {type: string}}
        - {name: subcategory, in: query, schema: {type: string}}
        - {name: difficulty, in: query, schema: {type: string, enum: [easy, medium, hard]}}
        - {name: search, in: query, schema: {type: string, maxLength: 120}}
      responses:
        '200': {description: Element collection}
  /elements/{id}:
    get:
      summary: Get one workout element
      parameters:
        - {name: id, in: path, required: true, schema: {type: integer}}
      responses:
        '200': {description: Element}
        '404': {description: Not found}
  /elements/category/{cat}:
    get:
      summary: List elements by category
      parameters:
        - {name: cat, in: path, required: true, schema: {type: string}}
      responses:
        '200': {description: Element collection}
  /bars:
    get: {summary: Get equipment catalog, responses: {'200': {description: Equipment}}}
  /base:
    get: {summary: Get fundamentals, responses: {'200': {description: Fundamentals}}}
  /safety:
    get: {summary: Get safety guidance, responses: {'200': {description: Safety}}}
  /stats:
    get: {summary: Get catalog statistics, responses: {'200': {description: Statistics}}}
  /metrics:
    get: {summary: Runtime metrics, responses: {'200': {description: Metrics}}}

  /auth/register:
    post:
      summary: Create account
      responses:
        '200': {description: Account created}
        '409': {description: Email already registered}
  /auth/login:
    post:
      summary: Create session
      responses:
        '200': {description: Authenticated}
        '401': {description: Invalid credentials}
  /auth/me:
    get:
      summary: Current account
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Account}
  /auth/logout:
    post:
      summary: End current session
      security: [{bearerAuth: []}]
      responses:
        '204': {description: Session ended}
  /progress:
    get: {summary: Get user progress, security: [{bearerAuth: []}], responses: {'200': {description: Progress}}}
    put: {summary: Update user progress, security: [{bearerAuth: []}], responses: {'200': {description: Updated}}}
  /workouts:
    post: {summary: Create workout session, security: [{bearerAuth: []}], responses: {'200': {description: Session created}}}
  /workouts/{id}:
    post:
      summary: Complete workout session
      security: [{bearerAuth: []}]
      parameters: [{name: id, in: path, required: true, schema: {type: string, format: uuid}}]
      responses: {'200': {description: Completed session}}
  /history:
    get: {summary: Workout history, security: [{bearerAuth: []}], responses: {'200': {description: History}}}
  /analytics:
    get: {summary: Training analytics, security: [{bearerAuth: []}], responses: {'200': {description: Analytics}}}
  /skills:
    get: {summary: Skill tree progress, security: [{bearerAuth: []}], responses: {'200': {description: Skill progress}}}
    put: {summary: Update skill level, security: [{bearerAuth: []}], responses: {'200': {description: Updated}}}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer